+++ client/src/features/category/category.ts
@@ -1,4 +1,4 @@
-import { Deletable } from "../../components/DeleteItem";
+import { Deletable } from "../../components/deleteItem.h";
 
 export interface Category extends Deletable {
   id: string;
