+++ client/src/features/questions/QuestionCard.tsx
@@ -1,16 +1,6 @@
 import React from "react";
 import { Grid, Paper, makeStyles, Typography } from "@material-ui/core";
-import { Deletable } from "../../components/DeleteItem";
-
-export interface MultipleChoiceQuestion extends Deletable {
-  category: string;
-  correct_answer: string;
-  difficulty: string;
-  id: string;
-  incorrect_answers: string[];
-  question: string;
-  type: string;
-}
+import { MultipleChoiceQuestion } from "./multiplechoice";
 
 type QuestionComponent = {
   question: MultipleChoiceQuestion;
