+++ client/src/features/questions/multiplechoice.ts
@@ -1,0 +1,11 @@
+import { Deletable } from "../../components/deleteItem.h";
+
+export interface MultipleChoiceQuestion extends Deletable {
+  category: string;
+  correct_answer: string;
+  difficulty: string;
+  id: string;
+  incorrect_answers: string[];
+  question: string;
+  type: string;
+}
